Implements some methods from PaymentService, changes in xml request for MakePayment Operation

+++ src/main/java/com/kyc/payments/entity/PaymentEntity.java
@@ -23,7 +23,7 @@ public class PaymentEntity {
 
     @Id
     @Column(name = "FOLIO")
-    private Integer folio;
+    private Long folio;
 
     @Column(name = "PAYMENT_SOURCE")
     private String paymentSource;
